requirements.txt
timm
yacs
opencv-python
matplotlib
pillow
scikit-learn
seaborn
scipy
tqdm
triton
PyYAML
PyWavelets
pandas
mat4py
einops
faiss-gpu
sympy
ftfy
regex
termcolor
tabulate
fvcore